case/mokuai/requirements.txt
aes==1.0.0
allure-pytest==2.8.7
allure-python-commons==2.8.7
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
crypto==1.4.1
cryptography==2.8
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
Jinja2==2.11.1
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.1.0
Naked==0.1.31
packaging==20.0
Padding==0.5
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.2
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-repeat==0.8.0
PyYAML==5.3
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
shellescape==3.8.1
six==1.13.0
Steganocrypto==0.0.3
urllib3==1.25.7
wcwidth==0.1.8
